+++ src/main/java/com/test/carina/demo/api/todos/DeleteTodoMethod.java
@@ -1,4 +1,4 @@
-package com.test.carina.demo.api.todos;
+package api.todos;
 
 import com.qaprosoft.carina.core.foundation.api.AbstractApiMethodV2;
 import com.qaprosoft.carina.core.foundation.api.annotation.Endpoint;
@@ -9,13 +9,12 @@ import com.qaprosoft.carina.core.foundation.api.http.HttpMethodType;
 import com.qaprosoft.carina.core.foundation.api.http.HttpResponseStatusType;
 import com.zebrunner.carina.utils.Configuration;
 
-@Endpoint(url = "${base_url}/todos", methodType = HttpMethodType.POST)
-@RequestTemplatePath(path = "api/todos/_post/rq.json")
-@ResponseTemplatePath(path = "api/todos/_post/rs.json")
+@Endpoint(url = "${base_url}/todos/1", methodType = HttpMethodType.DELETE)
+@RequestTemplatePath(path = "api/todos/_delete/rq.json")
+@ResponseTemplatePath(path = "api/todos/_delete/rs.json")
 @SuccessfulHttpStatus(status = HttpResponseStatusType.OK_200)
-public class PostTodoMethod extends AbstractApiMethodV2 {
-    public PostTodoMethod() {
-        super("api/todos/_post/rq.json", "api/todos/_post/rs.json");
+public class DeleteTodoMethod extends AbstractApiMethodV2 {
+    public DeleteTodoMethod() {
         replaceUrlPlaceholder("base_url", Configuration.getEnvArg("api_url"));
     }
