+++ src/main/java/com/test/carina/demo/api/todos/GetAllTodoMethod.java
@@ -1,21 +1,18 @@
-package com.test.carina.demo.api.todos;
+package api.todos;
 
 import com.qaprosoft.carina.core.foundation.api.AbstractApiMethodV2;
 import com.qaprosoft.carina.core.foundation.api.annotation.Endpoint;
-import com.qaprosoft.carina.core.foundation.api.annotation.RequestTemplatePath;
 import com.qaprosoft.carina.core.foundation.api.annotation.ResponseTemplatePath;
 import com.qaprosoft.carina.core.foundation.api.annotation.SuccessfulHttpStatus;
 import com.qaprosoft.carina.core.foundation.api.http.HttpMethodType;
 import com.qaprosoft.carina.core.foundation.api.http.HttpResponseStatusType;
 import com.zebrunner.carina.utils.Configuration;
 
-@Endpoint(url = "${base_url}/todos/1", methodType = HttpMethodType.DELETE)
-@RequestTemplatePath(path = "api/todos/_delete/rq.json")
-@ResponseTemplatePath(path = "api/todos/_delete/rs.json")
+@Endpoint(url = "${base_url}/todos", methodType = HttpMethodType.GET)
+@ResponseTemplatePath(path = "api/todos/_getall/rs.json")
 @SuccessfulHttpStatus(status = HttpResponseStatusType.OK_200)
-public class DeleteTodoMethod extends AbstractApiMethodV2 {
-    public DeleteTodoMethod() {
+public class GetAllTodoMethod extends AbstractApiMethodV2 {
+    public GetAllTodoMethod() {
         replaceUrlPlaceholder("base_url", Configuration.getEnvArg("api_url"));
     }
-
 }
