+++ src/main/java/com/test/carina/demo/api/todos/PostTodoMethod.java
@@ -1,18 +1,22 @@
-package com.test.carina.demo.api.todos;
+package api.todos;
 
 import com.qaprosoft.carina.core.foundation.api.AbstractApiMethodV2;
 import com.qaprosoft.carina.core.foundation.api.annotation.Endpoint;
+import com.qaprosoft.carina.core.foundation.api.annotation.RequestTemplatePath;
 import com.qaprosoft.carina.core.foundation.api.annotation.ResponseTemplatePath;
 import com.qaprosoft.carina.core.foundation.api.annotation.SuccessfulHttpStatus;
 import com.qaprosoft.carina.core.foundation.api.http.HttpMethodType;
 import com.qaprosoft.carina.core.foundation.api.http.HttpResponseStatusType;
 import com.zebrunner.carina.utils.Configuration;
 
-@Endpoint(url = "${base_url}/todos", methodType = HttpMethodType.GET)
-@ResponseTemplatePath(path = "api/todos/_getall/rs.json")
+@Endpoint(url = "${base_url}/todos", methodType = HttpMethodType.POST)
+@RequestTemplatePath(path = "api/todos/_post/rq.json")
+@ResponseTemplatePath(path = "api/todos/_post/rs.json")
 @SuccessfulHttpStatus(status = HttpResponseStatusType.OK_200)
-public class GetAllTodoMethod extends AbstractApiMethodV2 {
-    public GetAllTodoMethod(){
+public class PostTodoMethod extends AbstractApiMethodV2 {
+    public PostTodoMethod() {
+        super("api/todos/_post/rq.json", "api/todos/_post/rs.json");
         replaceUrlPlaceholder("base_url", Configuration.getEnvArg("api_url"));
     }
+
 }
