+++ src/main/java/resources/_config.properties
@@ -1,7 +1,6 @@
 #=====================================================#
 #================= Configuration v2 ==================#
 #=====================================================#
-
 #=================== Global configuration ============#
 url=https://demoblaze.com
 capabilities.browserName=chrome
@@ -10,41 +9,34 @@ selenium_url=http://localhost:4444/wd/hub
 custom_capabilities=NULL
 extra_capabilities=NULL
 app_version=
-
 #desired thread count for parallel execution. If commented "thread-count" value from TestNG suite xml is used.
 #thread_count=1
-
 #desired thread count for data provider(s). If commented "data-provider-thread-count" value from TestNG suite xml is used.
 #data_provider_thread_count=1
 #=====================================================#
-
 #=============== Environment configuration ===========#
 env=DEMO
 DEMO.base=https://www.gsmarena.com
 DEMO.api_url=https://jsonplaceholder.typicode.com
-STAG.api_url=
-DEMO.soap_url=https://www.crcind.com:443/csp/samples/SOAP.Demo.cls
+#STAG.api_url=
+#DEMO.soap_url=https://www.crcind.com:443/csp/samples/SOAP.Demo.cls
 #=====================================================#
-
 #============== Locale Configurations ================#
 locale=en_US
 localization_testing=false
 localization_encoding=UTF-8
 language=en_US
 #=====================================================#
-
 #================ Retry configuration ================#
 #number of extra retries to execute tests until pass
 retry_count=0
 #=====================================================#
-
 #============== WebDriver configuration ==============#
 explicit_timeout=20
 #number of extra retries to start driver session and interval in seconds
 init_retry_count=0
 init_retry_interval=1
 #=====================================================#
-
 #uncomment and change udid and deviceName for Android test
 #============ Android Local Mobile ===================#
 #selenium_url=http://localhost:4723/wd/hub
@@ -55,7 +47,6 @@ init_retry_interval=1
 #capabilities.platformName=ANDROID
 #capabilities.app=https://qaprosoft.s3-us-west-2.amazonaws.com/carinademoexample.apk
 #=====================================================#
-
 #uncomment and change deviceName and platformVersion for iOS test
 #======== Local Run for iOS Mobile ===============#
 #selenium_url=http://localhost:4723/wd/hub
@@ -67,41 +58,34 @@ init_retry_interval=1
 #capabilities.udid=
 #capabilities.app=https://qaprosoft.s3-us-west-2.amazonaws.com/carinademoexample.zip
 #=====================================================#
-
 #================ Report configuration ===============#
 core_log_level=INFO
 auto_screenshot=true
 project_report_directory=./reports
 report_url=NULL
 max_screen_history=10
-
 email_list=NULL
 suite_name=NULL
 #=====================================================#
-
 #=============== MyBatis configuration ===============#
 mybatis_cnfg=mybatis.config.xml
 mybatis_env=testing
 #=====================================================#
-
 #====================== TestRail =====================#
 testrail_milestone=NULL
 testrail_assignee=NULL
 #=====================================================#
-
 #===================== Amazon ========================#
 s3_bucket_name=NULL
 access_key_id=NULL
 secret_key=NULL
 #=====================================================#
-
 #===================== Azure =========================#
-azure_account_name=NULL
-azure_blob_url=https://${azure_account_name}.blob.core.windows.net
-azure_access_key_token=NULL
-azure_local_storage=NULL
+#azure_account_name=NULL
+#azure_blob_url=https://${azure_account_name}.blob.core.windows.net
+#azure_access_key_token=NULL
+#azure_local_storage=NULL
 #=====================================================#
-
 #===================== Crypto ========================#
 crypto_key_value=OIujpEmIVZ0C9kOkXniFRw==
 #=====================================================#
