--- src/main/java/resources/_database.properties
@@ -1,7 +1,0 @@
-#===============================================================#
-#================== Database configuration ====================#
-#===============================================================#
-db.url=jdbc:postgresql://localhost:5432/postgres
-db.driver=org.postgresql.Driver
-db.user=postgres
-db.pass=postgres
