--- src/main/java/resources/_email.properties
@@ -1,13 +1,0 @@
-#===============================================================#
-#================== Email base configuration ===================#
-#===============================================================#
-
-title=Carina Demo Tests
-
-#================= SMTP server configuration ==================#
-mail.smtp.host=smtp.gmail.com
-mail.smtp.socketFactory.port=465
-mail.smtp.socketFactory.class=javax.net.ssl.SSLSocketFactory
-mail.smtp.auth=true
-mail.smtp.port=465
-#==============================================================#
