+++ src/main/java/resources/_testdata.properties
@@ -1,5 +1,6 @@
 #===============================================================#
 #========================== Test data ==========================#
 #===============================================================#
-
 test_credentials=test@gmail.com/{crypt:8O9iA4+f3nMzz85szmvKmQ==}
+username=username
+password=password
