--- src/test/resources/agent.properties
@@ -1,5 +1,0 @@
-reporting.enabled=false
-reporting.server.hostname=
-reporting.server.access-token=
-reporting.project-key=DEF
-reporting.run.environment=
