+++ src/test/resources/api/soap/soap.properties
@@ -2,7 +2,6 @@
 firstNumber=4
 secondNumber=6
 result=10
-
 #=============== LookupCity properties ===========#
 zipCode=10005
 city=New York
