+++ src/test/resources/api/todos/_delete/rq.json
@@ -3,12 +3,12 @@
     "userId": 1,
     "id": 1,
     "title": "delectus aut autem",
-    "completed":false
+    "completed": false
   },
   {
     "userId": 2,
     "id": 2,
     "title": "delectus aut autem",
-    "completed":false
+    "completed": false
   }
 ]
