+++ src/test/resources/api/users/_post/rq.json
@@ -1,8 +1,28 @@
 [
   {
     "id": 1,
-    <#if name??>"name": "${name}",</#if>
-    <#if username??>"username": "${username}",</#if>
+    <
+    #
+    if
+    name
+    ?
+    ?
+    >
+    "name": "${name}",
+    </
+    #
+    if>
+    <
+    #
+    if
+    username
+    ?
+    ?
+    >
+    "username": "${username}",
+    </
+    #
+    if>
     "email": "Sincere@april.biz",
     "address": {
       "street": "Kulas Light",
