+++ src/test/resources/testng_suites/demoblaze.xml
@@ -1,0 +1,23 @@
+<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
+<suite verbose="1" name="Carina Demo Tests - Web Single Driver Sample" parallel="tests">
+
+    <parameter name="suiteOwner" value="qpsdemo"/>
+    <parameter name="jenkinsJobName" value="Web-Demo-Single-Driver-Test"/>
+    <parameter name="jenkinsJobType" value="web"/>
+    <parameter name="jenkinsEnvironments" value="DEMO"/>
+    <parameter name="jenkinsEmail" value="demo@qaprosoft.com"/>
+
+    <parameter name="jenkinsRegressionPipeline" value="web_regression, full_regression"/>
+
+    <parameter name="jenkinsRegressionMatrix_full_regression" value="browser: chrome; browser: firefox"/>
+    <parameter name="jenkinsJobExecutionOrder" value="2"/>
+    <parameter name="jenkinsJobExecutionMode" value="continue"/>
+    <parameter name="jenkinsAutoScreenshot" value="true"/>
+
+    <test name="GSM arena web tests using single driver between tests">
+        <classes>
+            <class name="com.test.carina.demo.DemoBlazeTest" />
+        </classes>
+    </test>
+
+</suite>
